refactor logic into hook
